+++ masci_tools/io/fleurxmlmodifier.py
@@ -28,14 +28,15 @@ from masci_tools.io.io_fleurxml import load_inpxml
 from masci_tools.util.typing import XMLFileLike, FileLike
 from pathlib import Path
 from lxml import etree
+import warnings
 #Enable warnings for missing docstrings
 #pylint: enable=missing-function-docstring
 
 
 class ModifierTask(NamedTuple):
     name: str
-    args: tuple[Any, ...]
-    kwargs: dict[str, Any]
+    args: tuple[Any, ...] = ()
+    kwargs: dict[str, Any] = {}
 
 
 class FleurXMLModifier:
@@ -129,6 +130,15 @@ class FleurXMLModifier:
                 raise ValueError(f"Unknown modification method '{name}'") from exc
 
     def _validate_signature(self, name: str, *args: Any, **kwargs: Any) -> None:
+        """
+        DEPRECATED: use `_validate_arguments` instead without unpacking args/kwargs
+        """
+        warnings.warn(
+            'The _validate_signature method is deprecated. '
+            'Please use _validate_arguments without unpacking args/kwargs instead', DeprecationWarning)
+        self._validate_arguments(name, args, kwargs)
+
+    def _validate_arguments(self, name: str, args: tuple[Any, ...], kwargs: dict[str, Any]) -> None:
         """
         Validate that the given arguments to the registration
         method can be used to call the corresponding XML modifying function
@@ -327,17 +337,20 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.set_inpchanges()` to
         the list of tasks that will be done on the xmltree.
 
-        :param change_dict: a dictionary with changes
+        :param changes: a dictionary with changes
         :param path_spec: dict, with ggf. necessary further specifications for the path of the attribute
 
-        An example of change_dict::
+        An example of changes::
 
-            change_dict = {'itmax' : 1,
-                           'l_noco': True,
-                           'ctail': False,
-                           'l_ss': True}
+            changes = {'itmax' : 1,
+                       'l_noco': True,
+                       'ctail': False,
+                       'l_ss': True}
         """
-        self._validate_signature('set_inpchanges', *args, **kwargs)
+        if 'change_dict' in kwargs:
+            warnings.warn('The argument change_dict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('change_dict')
+        self._validate_arguments('set_inpchanges', args, kwargs)
         self._tasks.append(ModifierTask('set_inpchanges', args, kwargs))
 
     def shift_value(self, *args: Any, **kwargs: Any) -> None:
@@ -345,15 +358,20 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.shift_value()` to
         the list of tasks that will be done on the xmltree.
 
-        :param change_dict: a python dictionary with the keys to shift and the shift values.
-        :param mode: 'abs' if change given is absolute, 'rel' if relative
+        :param changes: a python dictionary with the keys to shift and the shift values.
+        :param mode: str (either `rel`/`relative` or `abs`/`absolute`).
+                     `rel`/`relative` multiplies the old value with the given value
+                     `abs`/`absolute` adds the old value and the given value
         :param path_spec: dict, with ggf. necessary further specifications for the path of the attribute
 
-        An example of change_dict::
+        An example of changes::
 
-                change_dict = {'itmax' : 1, 'dVac': -0.123}
+                changes = {'itmax' : 1, 'dVac': -0.123}
         """
-        self._validate_signature('shift_value', *args, **kwargs)
+        if 'change_dict' in kwargs:
+            warnings.warn('The argument change_dict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('change_dict')
+        self._validate_arguments('shift_value', args, kwargs)
         self._tasks.append(ModifierTask('shift_value', args, kwargs))
 
     def set_species(self, *args: Any, **kwargs: Any) -> None:
@@ -363,27 +381,29 @@ class FleurXMLModifier:
 
         :param species_name: string, name of the specie you want to change
                              Can be name of the species, 'all' or 'all-<string>' (sets species with the string in the species name)
-        :param attributedict: a python dict specifying what you want to change.
+        :param changes: a python dict specifying what you want to change.
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
-        :param create: bool, if species does not exist create it and all subtags?
 
-        **attributedict** is a python dictionary containing dictionaries that specify attributes
+        **changes** is a python dictionary containing dictionaries that specify attributes
         to be set inside the certain specie. For example, if one wants to set a MT radius it
         can be done via::
 
-            attributedict = {'mtSphere' : {'radius' : 2.2}}
+            changes = {'mtSphere' : {'radius' : 2.2}}
 
         Another example::
 
-            'attributedict': {'special': {'socscale': 0.0}}
+            'changes': {'special': {'socscale': 0.0}}
 
         that switches SOC terms on a sertain specie. ``mtSphere``, ``atomicCutoffs``,
         ``energyParameters``, ``lo``, ``electronConfig``, ``nocoParams``, ``ldaU`` and
         ``special`` keys are supported. To find possible
         keys of the inner dictionary please refer to the FLEUR documentation flapw.de
         """
-        self._validate_signature('set_species', *args, **kwargs)
+        if 'attributedict' in kwargs:
+            warnings.warn('The argument attributedict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('attributedict')
+        self._validate_arguments('set_species', args, kwargs)
         self._tasks.append(ModifierTask('set_species', args, kwargs))
 
     def set_species_label(self, *args: Any, **kwargs: Any) -> None:
@@ -392,10 +412,13 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param atom_label: string, a label of the atom which specie will be changed. 'all' to change all the species
-        :param attributedict: a python dict specifying what you want to change.
+        :param changes: a python dict specifying what you want to change.
 
         """
-        self._validate_signature('set_species_label', *args, **kwargs)
+        if 'attributedict' in kwargs:
+            warnings.warn('The argument attributedict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('attributedict')
+        self._validate_arguments('set_species_label', args, kwargs)
         self._tasks.append(ModifierTask('set_species_label', args, kwargs))
 
     def clone_species(self, *args: Any, **kwargs: Any) -> None:
@@ -408,7 +431,7 @@ class FleurXMLModifier:
         :param new_name: new name of the cloned species
         :param changes: a optional python dict specifying what you want to change.
         """
-        self._validate_signature('clone_species', *args, **kwargs)
+        self._validate_arguments('clone_species', args, kwargs)
         self._tasks.append(ModifierTask('clone_species', args, kwargs))
 
     def switch_species(self, *args: Any, **kwargs: Any) -> None:
@@ -425,7 +448,7 @@ class FleurXMLModifier:
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details`
         """
-        self._validate_signature('switch_species', *args, **kwargs)
+        self._validate_arguments('switch_species', args, kwargs)
         self._tasks.append(ModifierTask('switch_species', args, kwargs))
 
     def switch_species_label(self, *args: Any, **kwargs: Any) -> None:
@@ -439,7 +462,7 @@ class FleurXMLModifier:
                   from one species the species will be cloned with :py:func:`clone_species()`
         :param changes: changes to do if the species is cloned
         """
-        self._validate_signature('switch_species_label', *args, **kwargs)
+        self._validate_arguments('switch_species_label', args, kwargs)
         self._tasks.append(ModifierTask('switch_species_label', args, kwargs))
 
     def shift_value_species_label(self, *args: Any, **kwargs: Any) -> None:
@@ -448,16 +471,24 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param atom_label: string, a label of the atom which specie will be changed. 'all' if set up all species
-        :param attributename: name of the attribute to change
-        :param value_given: value to add or to multiply by
-        :param mode: 'rel' for multiplication or 'abs' for addition
+        :param attribute_name: name of the attribute to change
+        :param number_to_add: value to add or to multiply by
+        :param mode: str (either `rel`/`relative` or `abs`/`absolute`).
+                    `rel`/`relative` multiplies the old value with `number_to_add`
+                    `abs`/`absolute` adds the old value and `number_to_add`
 
-        Kwargs if the attributename does not correspond to a unique path:
+        Kwargs if the attribute_name does not correspond to a unique path:
             :param contains: str, this string has to be in the final path
             :param not_contains: str, this string has to NOT be in the final path
 
         """
-        self._validate_signature('shift_value_species_label', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use attribute_name instead', DeprecationWarning)
+            kwargs['attribute_name'] = kwargs.pop('attributename')
+        if 'value_given' in kwargs:
+            warnings.warn('The argument value_given is deprecated. Use number_to_add instead', DeprecationWarning)
+            kwargs['number_to_add'] = kwargs.pop('value_given')
+        self._validate_arguments('shift_value_species_label', args, kwargs)
         self._tasks.append(ModifierTask('shift_value_species_label', args, kwargs))
 
     def set_atomgroup(self, *args: Any, **kwargs: Any) -> None:
@@ -465,21 +496,26 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.set_atomgroup()` to
         the list of tasks that will be done on the xmltree.
 
-        :param attributedict: a python dict specifying what you want to change.
+        :param changes: a python dict specifying what you want to change.
         :param position: position of an atom group to be changed. If equals to 'all', all species will be changed
         :param species: atom groups, corresponding to the given species will be changed
-        :param create: bool, if species does not exist create it and all subtags?
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
 
-        **attributedict** is a python dictionary containing dictionaries that specify attributes
+        **changes** is a python dictionary containing dictionaries that specify attributes
         to be set inside the certain specie. For example, if one wants to set a beta noco parameter it
         can be done via::
 
-            'attributedict': {'nocoParams': {'beta': val}}
+            'changes': {'nocoParams': {'beta': val}}
 
         """
-        self._validate_signature('set_atomgroup', *args, **kwargs)
+        if 'attributedict' in kwargs:
+            warnings.warn('The argument attributedict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('attributedict')
+        if 'create' in kwargs:
+            warnings.warn('The argument create is deprecated and is ignored.', DeprecationWarning)
+            kwargs.pop('create')
+        self._validate_arguments('set_atomgroup', args, kwargs)
         self._tasks.append(ModifierTask('set_atomgroup', args, kwargs))
 
     def set_atomgroup_label(self, *args: Any, **kwargs: Any) -> None:
@@ -488,17 +524,22 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param atom_label: string, a label of the atom which specie will be changed. 'all' to change all the species
-        :param attributedict: a python dict specifying what you want to change.
-        :param create: bool, if species does not exist create it and all subtags?
+        :param changes: a python dict specifying what you want to change.
 
-        **attributedict** is a python dictionary containing dictionaries that specify attributes
+        **changes** is a python dictionary containing dictionaries that specify attributes
         to be set inside the certain specie. For example, if one wants to set a beta noco parameter it
         can be done via::
 
-            'attributedict': {'nocoParams': {'beta': val}}
+            'changes': {'nocoParams': {'beta': val}}
 
         """
-        self._validate_signature('set_atomgroup_label', *args, **kwargs)
+        if 'attributedict' in kwargs:
+            warnings.warn('The argument attributedict is deprecated. Use changes instead', DeprecationWarning)
+            kwargs['changes'] = kwargs.pop('attributedict')
+        if 'create' in kwargs:
+            warnings.warn('The argument create is deprecatedand is ignored.', DeprecationWarning)
+            kwargs.pop('create')
+        self._validate_arguments('set_atomgroup_label', args, kwargs)
         self._tasks.append(ModifierTask('set_atomgroup_label', args, kwargs))
 
     def create_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -520,7 +561,7 @@ class FleurXMLModifier:
             :param contains: str, this string has to be in the final path
             :param not_contains: str, this string has to NOT be in the final path
         """
-        self._validate_signature('create_tag', *args, **kwargs)
+        self._validate_arguments('create_tag', args, kwargs)
         self._tasks.append(ModifierTask('create_tag', args, kwargs))
 
     def delete_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -540,7 +581,7 @@ class FleurXMLModifier:
             :param contains: str, this string has to be in the final path
             :param not_contains: str, this string has to NOT be in the final path
         """
-        self._validate_signature('delete_tag', *args, **kwargs)
+        self._validate_arguments('delete_tag', args, kwargs)
         self._tasks.append(ModifierTask('delete_tag', args, kwargs))
 
     def delete_att(self, *args: Any, **kwargs: Any) -> None:
@@ -548,7 +589,7 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.delete_att()` to
         the list of tasks that will be done on the xmltree.
 
-        :param tag: str of the attribute to delete
+        :param name: str of the attribute to delete
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
         :param occurrences: int or list of int. Which occurrence of the parent nodes to delete a attribute.
                             By default all nodes are used.
@@ -563,7 +604,11 @@ class FleurXMLModifier:
             :param exclude: list of str, here specific types of attributes can be excluded
                             valid values are: settable, settable_contains, other
         """
-        self._validate_signature('delete_att', *args, **kwargs)
+        if 'attrib_name' in kwargs:
+            warnings.warn('The argument attrib_name is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attrib_name')
+        print(args, kwargs)
+        self._validate_arguments('delete_att', args, kwargs)
         self._tasks.append(ModifierTask('delete_att', args, kwargs))
 
     def replace_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -572,7 +617,7 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param tag: str of the tag to replace
-        :param newelement: a new tag
+        :param element: a new tag
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
         :param occurrences: int or list of int. Which occurrence of the parent nodes to replace a tag.
                             By default all nodes are used.
@@ -584,7 +629,10 @@ class FleurXMLModifier:
             :param contains: str, this string has to be in the final path
             :param not_contains: str, this string has to NOT be in the final path
         """
-        self._validate_signature('replace_tag', *args, **kwargs)
+        if 'newelement' in kwargs:
+            warnings.warn('The argument newelement is deprecated. Use element instead', DeprecationWarning)
+            kwargs['element'] = kwargs.pop('newelement')
+        self._validate_arguments('replace_tag', args, kwargs)
         self._tasks.append(ModifierTask('replace_tag', args, kwargs))
 
     def set_complex_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -593,8 +641,8 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param tag_name: name of the tag to set
-        :param attributedict: Keys in the dictionary correspond to names of tags and the values are the modifications
-                              to do on this tag (attributename, subdict with changes to the subtag, ...)
+        :param changes: Keys in the dictionary correspond to names of tags and the values are the modifications
+                        to do on this tag (attributename, subdict with changes to the subtag, ...)
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
         :param create: bool optional (default False), if True and the path, where the complex tag is
                        set does not exist it is created
@@ -607,7 +655,7 @@ class FleurXMLModifier:
             :param not_contains: str, this string has to NOT be in the final path
 
         """
-        self._validate_signature('set_complex_tag', *args, **kwargs)
+        self._validate_arguments('set_complex_tag', args, kwargs)
         self._tasks.append(ModifierTask('set_complex_tag', args, kwargs))
 
     def set_simple_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -629,7 +677,7 @@ class FleurXMLModifier:
             :param contains: str, this string has to be in the final path
             :param not_contains: str, this string has to NOT be in the final path
         """
-        self._validate_signature('set_simple_tag', *args, **kwargs)
+        self._validate_arguments('set_simple_tag', args, kwargs)
         self._tasks.append(ModifierTask('set_simple_tag', args, kwargs))
 
     def set_text(self, *args: Any, **kwargs: Any) -> None:
@@ -651,7 +699,7 @@ class FleurXMLModifier:
             :param not_contains: str, this string has to NOT be in the final path
 
         """
-        self._validate_signature('set_text', *args, **kwargs)
+        self._validate_arguments('set_text', args, kwargs)
         self._tasks.append(ModifierTask('set_text', args, kwargs))
 
     def set_first_text(self, *args: Any, **kwargs: Any) -> None:
@@ -672,7 +720,7 @@ class FleurXMLModifier:
             :param not_contains: str, this string has to NOT be in the final path
 
         """
-        self._validate_signature('set_first_text', *args, **kwargs)
+        self._validate_arguments('set_first_text', args, kwargs)
         self._tasks.append(ModifierTask('set_first_text', args, kwargs))
 
     def set_attrib_value(self, *args: Any, **kwargs: Any) -> None:
@@ -680,8 +728,8 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.set_attrib_value()` to
         the list of tasks that will be done on the xmltree.
 
-        :param attributename: the attribute name to set
-        :param attribv: value or list of values to set
+        :param name: the attribute name to set
+        :param value: value or list of values to set
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
         :param occurrences: int or list of int. Which occurrence of the node to set. By default all are set.
         :param create: bool optional (default False), if True the tag is created if is missing
@@ -697,7 +745,13 @@ class FleurXMLModifier:
                             valid values are: settable, settable_contains, other
 
         """
-        self._validate_signature('set_attrib_value', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        if 'attribv' in kwargs:
+            warnings.warn('The argument attribv is deprecated. Use value instead', DeprecationWarning)
+            kwargs['value'] = kwargs.pop('attribv')
+        self._validate_arguments('set_attrib_value', args, kwargs)
         self._tasks.append(ModifierTask('set_attrib_value', args, kwargs))
 
     def set_first_attrib_value(self, *args: Any, **kwargs: Any) -> None:
@@ -705,8 +759,8 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.set_first_attrib_value()` to
         the list of tasks that will be done on the xmltree.
 
-        :param attributename: the attribute name to set
-        :param attribv: value or list of values to set
+        :param name: the attribute name to set
+        :param value: value or list of values to set
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
         :param create: bool optional (default False), if True the tag is created if is missing
         :param filters: Dict specifying constraints to apply on the xpath.
@@ -721,7 +775,13 @@ class FleurXMLModifier:
                             valid values are: settable, settable_contains, other
 
         """
-        self._validate_signature('set_first_attrib_value', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        if 'attribv' in kwargs:
+            warnings.warn('The argument attribv is deprecated. Use value instead', DeprecationWarning)
+            kwargs['value'] = kwargs.pop('attribv')
+        self._validate_arguments('set_first_attrib_value', args, kwargs)
         self._tasks.append(ModifierTask('set_first_attrib_value', args, kwargs))
 
     def add_number_to_attrib(self, *args: Any, **kwargs: Any) -> None:
@@ -729,12 +789,12 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.add_number_to_attrib()` to
         the list of tasks that will be done on the xmltree.
 
-        :param attributename: the attribute name to change
-        :param add_number: number to add/multiply with the old attribute value
+        :param name: the attribute name to change
+        :param number_to_add: number to add/multiply with the old attribute value
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
-        :param mode: str (either `rel` or `abs`).
-                     `rel` multiplies the old value with `add_number`
-                     `abs` adds the old value and `add_number`
+        :param mode: str (either `rel`/`relative` or `abs`/`absolute`).
+                    `rel`/`relative` multiplies the old value with `number_to_add`
+                    `abs`/`absolute` adds the old value and `number_to_add`
         :param occurrences: int or list of int. Which occurrence of the node to set. By default all are set.
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
@@ -748,7 +808,13 @@ class FleurXMLModifier:
                             valid values are: settable, settable_contains, other
 
         """
-        self._validate_signature('add_number_to_attrib', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        if 'add_number' in kwargs:
+            warnings.warn('The argument add_number is deprecated. Use number_to_add instead', DeprecationWarning)
+            kwargs['number_to_add'] = kwargs.pop('add_number')
+        self._validate_arguments('add_number_to_attrib', args, kwargs)
         self._tasks.append(ModifierTask('add_number_to_attrib', args, kwargs))
 
     def add_number_to_first_attrib(self, *args: Any, **kwargs: Any) -> None:
@@ -756,12 +822,12 @@ class FleurXMLModifier:
         Appends a :py:func:`~masci_tools.util.xml.xml_setters_names.add_number_to_first_attrib()` to
         the list of tasks that will be done on the xmltree.
 
-        :param attributename: the attribute name to change
-        :param add_number: number to add/multiply with the old attribute value
+        :param name: the attribute name to change
+        :param number_to_add: number to add/multiply with the old attribute value
         :param complex_xpath: an optional xpath to use instead of the simple xpath for the evaluation
-        :param mode: str (either `rel` or `abs`).
-                     `rel` multiplies the old value with `add_number`
-                     `abs` adds the old value and `add_number`
+        :param mode: str (either `rel`/`relative` or `abs`/`absolute`).
+                    `rel`/`relative` multiplies the old value with `number_to_add`
+                    `abs`/`absolute` adds the old value and `number_to_add`
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
 
@@ -774,7 +840,13 @@ class FleurXMLModifier:
                             valid values are: settable, settable_contains, other
 
         """
-        self._validate_signature('add_number_to_first_attrib', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        if 'add_number' in kwargs:
+            warnings.warn('The argument add_number is deprecated. Use number_to_add instead', DeprecationWarning)
+            kwargs['number_to_add'] = kwargs.pop('add_number')
+        self._validate_arguments('add_number_to_first_attrib', args, kwargs)
         self._tasks.append(ModifierTask('add_number_to_first_attrib', args, kwargs))
 
     def xml_create_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -789,7 +861,7 @@ class FleurXMLModifier:
         :param occurrences: int or list of int. Which occurrence of the parent nodes to create a tag.
                             By default all nodes are used.
         """
-        self._validate_signature('xml_create_tag', *args, **kwargs)
+        self._validate_arguments('xml_create_tag', args, kwargs)
         self._tasks.append(ModifierTask('xml_create_tag', args, kwargs))
 
     def xml_replace_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -798,11 +870,15 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param xpath: a path to the tag to be replaced
-        :param newelement: a new tag
+        :param element: a new tag
         :param occurrences: int or list of int. Which occurrence of the parent nodes to create a tag.
                             By default all nodes are used.
         """
-        self._validate_signature('xml_replace_tag', *args, **kwargs)
+        if 'newelement' in kwargs:
+            warnings.warn('The argument newelement is deprecated. Use element instead', DeprecationWarning)
+            kwargs['element'] = kwargs.pop('newelement')
+
+        self._validate_arguments('xml_replace_tag', args, kwargs)
         self._tasks.append(ModifierTask('xml_replace_tag', args, kwargs))
 
     def xml_delete_tag(self, *args: Any, **kwargs: Any) -> None:
@@ -814,7 +890,7 @@ class FleurXMLModifier:
         :param occurrences: int or list of int. Which occurrence of the parent nodes to create a tag.
                             By default all nodes are used.
         """
-        self._validate_signature('xml_delete_tag', *args, **kwargs)
+        self._validate_arguments('xml_delete_tag', args, kwargs)
         self._tasks.append(ModifierTask('xml_delete_tag', args, kwargs))
 
     def xml_delete_att(self, *args: Any, **kwargs: Any) -> None:
@@ -823,11 +899,14 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param xpath: a path to the attribute to be deleted
-        :param attrib: the name of an attribute
+        :param name: the name of an attribute
         :param occurrences: int or list of int. Which occurrence of the parent nodes to create a tag.
                             By default all nodes are used.
         """
-        self._validate_signature('xml_delete_att', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        self._validate_arguments('xml_delete_att', args, kwargs)
         self._tasks.append(ModifierTask('xml_delete_att', args, kwargs))
 
     def xml_set_attrib_value_no_create(self, *args: Any, **kwargs: Any) -> None:
@@ -836,11 +915,17 @@ class FleurXMLModifier:
         the list of tasks that will be done on the xmltree.
 
         :param xpath: a path where to set the attributes
-        :param attributename: the attribute name to set
-        :param attribv: value or list of values to set (if not str they will be converted with `str(value)`)
+        :param name: the attribute name to set
+        :param value: value or list of values to set (if not str they will be converted with `str(value)`)
         :param occurrences: int or list of int. Which occurrence of the node to set. By default all are set.
         """
-        self._validate_signature('xml_set_attrib_value_no_create', *args, **kwargs)
+        if 'attributename' in kwargs:
+            warnings.warn('The argument attributename is deprecated. Use name instead', DeprecationWarning)
+            kwargs['name'] = kwargs.pop('attributename')
+        if 'attribv' in kwargs:
+            warnings.warn('The argument attribv is deprecated. Use value instead', DeprecationWarning)
+            kwargs['value'] = kwargs.pop('attribv')
+        self._validate_arguments('xml_set_attrib_value_no_create', args, kwargs)
         self._tasks.append(ModifierTask('xml_set_attrib_value_no_create', args, kwargs))
 
     def xml_set_text_no_create(self, *args: Any, **kwargs: Any) -> None:
@@ -852,7 +937,7 @@ class FleurXMLModifier:
         :param text: value or list of values to set (if not str they will be converted with `str(value)`)
         :param occurrences: int or list of int. Which occurrence of the node to set. By default all are set.
         """
-        self._validate_signature('xml_set_text_no_create', *args, **kwargs)
+        self._validate_arguments('xml_set_text_no_create', args, kwargs)
         self._tasks.append(ModifierTask('xml_set_text_no_create', args, kwargs))
 
     def set_nmmpmat(self, *args: Any, **kwargs: Any) -> None:
@@ -871,7 +956,7 @@ class FleurXMLModifier:
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
         """
-        self._validate_signature('set_nmmpmat', *args, **kwargs)
+        self._validate_arguments('set_nmmpmat', args, kwargs)
         self._tasks.append(ModifierTask('set_nmmpmat', args, kwargs))
 
     def rotate_nmmpmat(self, *args: Any, **kwargs: Any) -> None:
@@ -886,7 +971,7 @@ class FleurXMLModifier:
         :param filters: Dict specifying constraints to apply on the xpath.
                         See :py:class:`~masci_tools.util.xml.xpathbuilder.XPathBuilder` for details
         """
-        self._validate_signature('rotate_nmmpmat', *args, **kwargs)
+        self._validate_arguments('rotate_nmmpmat', args, kwargs)
         self._tasks.append(ModifierTask('rotate_nmmpmat', args, kwargs))
 
     def set_kpointlist(self, *args: Any, **kwargs: Any) -> None:
@@ -907,7 +992,7 @@ class FleurXMLModifier:
         :param switch: bool, if True the kPointlist will be used by Fleur when starting the next calculation
         :param overwrite: bool, if True and a kPointlist with the given name already exists it will be overwritten
         """
-        self._validate_signature('set_kpointlist', *args, **kwargs)
+        self._validate_arguments('set_kpointlist', args, kwargs)
         self._tasks.append(ModifierTask('set_kpointlist', args, kwargs))
 
     def switch_kpointset(self, *args: Any, **kwargs: Any) -> None:
@@ -920,7 +1005,7 @@ class FleurXMLModifier:
 
         :param list_name: name of the kPoint set to use
         """
-        self._validate_signature('switch_kpointset', *args, **kwargs)
+        self._validate_arguments('switch_kpointset', args, kwargs)
         self._tasks.append(ModifierTask('switch_kpointset', args, kwargs))
 
     def set_nkpts(self, *args: Any, **kwargs: Any) -> None:
@@ -935,7 +1020,7 @@ class FleurXMLModifier:
         :param gamma: bool that controls if the gamma-point should be included
                       in the k-point mesh
         """
-        self._validate_signature('set_nkpts', *args, **kwargs)
+        self._validate_arguments('set_nkpts', args, kwargs)
         self._tasks.append(ModifierTask('set_nkpts', args, kwargs))
 
     def set_kpath(self, *args: Any, **kwargs: Any) -> None:
@@ -951,5 +1036,5 @@ class FleurXMLModifier:
         :param gamma: bool that controls if the gamma-point should be included
                       in the k-point mesh
         """
-        self._validate_signature('set_kpath', *args, **kwargs)
+        self._validate_arguments('set_kpath', args, kwargs)
         self._tasks.append(ModifierTask('set_kpath', args, kwargs))